error cuenta corregido

+++ api/services/User.js
@@ -44,8 +44,10 @@ class UserServices {
     try {
       const data = await User.findByIdAndUpdate(id, body);
       data.image =`/${data.id}.jpg` ;
-      fs.rename("src/assets/img/users/01.jpg",`src/assets/img/users${data.image}`, err => {if (err) console.log(err)})  
-      //console.log(data)
+      if(fs.existsSync("src/assets/img/users/01.jpg")){
+        fs.rename("src/assets/img/users/01.jpg",`src/assets/img/users${data.image}`, err => {if (err) console.log(err)})
+      } 
+
       return {
         error: false,
         data: {
